% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_edge_attributes}
\alias{add_edge_attributes}
\title{Add edge attributes to a graph}
\usage{
add_edge_attributes(object, attr_name, value, edgelist, overwrite = FALSE)
}
\arguments{
\item{object}{graph}

\item{attr_name}{vector with one or more names of attributes}

\item{value}{vector, matrix, or data.frame}

\item{edgelist}{matrix or dataframe containing the edgelist and edge attributes}

\item{overwrite}{logical, should an edge attribute be overwritten if it already 
exists inside the graph?}
}
\value{
graph with the added edge attributes
}
\description{
Add one or more edge attributes to a graph
}
\details{
Consistent API to easily add one or more edge attributes to a graph of 
class \code{igraph} or \code{network}. Both the \code{igraph} and \code{statnet} 
packages do not have (simple) functions for this.

Adding edge attributes is easy, There are several options:
\itemize{
\item a vector with values specified in \code{value} and a name for the 
attribute specified in \code{attr_name}. 

This is only a very safe option if you know for sure that the order of the edges 
in \code{value} are identical to the order of the edges inside the graph object.!
\item a \code{data.frame} or \code{matrix} for \code{value}. 
Provide the names that the edge attributes should get in \code{attr_name} (which 
can be a vector with more than one name: the same number of names as there 
are columns in \code{value}). Note: any column names in \code{value} will be 
overridden by whatever is provided in \code{attr_name}.

This is only a very safe option if you know for sure that the order of the edges 
in \code{value} are identical to the order of the edges inside the graph object!
\item an edgelist (\code{data.frame} or \code{matrix}) in \code{edgelist}. 
If \code{attrname} is not provided, then all attributes in \code{edgelist} are 
added to the graph. 
If \code{attrname} is provided, only the columns in \code{edgelist} are added 
to the graph that occur in \code{attr_name}.

The edgelist is assumed to have at least three columns. The first column 
contains the sender and the second column contains the receiver of the edge 
(this is irrespective of the names of these columns and the 
sender-receiver order is irrelevant for undirected graphs). 
Any further columns are assumed to be edge attributes.

This is the safe option, as the function makes sure that the 
edge attributes are added to the correct edge.
}
}
