% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_graph_attribute}
\alias{add_graph_attribute}
\title{Add a graph attribute}
\usage{
add_graph_attribute(x, attr_name = NULL, value)
}
\arguments{
\item{x}{graph of class \code{igraph} or \code{network}}

\item{attr_name}{character, name to be used for the attribute}

\item{value}{the value to be given to the attribute}
}
\value{
graph with the added graph attribute
}
\description{
Add a graph-level attribute to the graph object
}
