% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_vertex_attributes}
\alias{add_vertex_attributes}
\title{Add vertex attributes to a graph}
\usage{
add_vertex_attributes(x, attr_name = NULL, value)
}
\arguments{
\item{x}{graph object to which the vertex attributes need to be added}

\item{attr_name}{optional, character vector with attribute names (see 'Details')}

\item{value}{\code{data.frame}, \code{matrix} (with column names), or a vector
(see 'Details')}
}
\value{
graph object with new attributes added
}
\description{
Add one or more vertex attributes to a graph
}
\details{
Consistent API to easily add one or more vertex attributes to a graph of 
class \code{igraph} or \code{network}. Both the \code{igraph} and \code{statnet} 
packages do not have (simple) functions for this.

Adding vertex attributes is easy, There are several options:
\itemize{
\item a vector with values specified in \code{value} and a name for the 
attribute specified in \code{attr_name}. 
\item a \code{data.frame} for the \code{value}. All columns in the \code{data.frame} 
will be added to the graph as vertex attributes and the names in the 
\code{data.frame} are used as the names of 
the new attributes,
\item a \code{matrix} for the \code{value}. The matrix needs to have appropriate 
column names. All columns in the \code{matrix} will be added to the graph as 
vertex attributes and the column names in the \code{matrix} are used as the names of 
the new attributes,
\item a \code{data.frame} or \code{matrix} (with column names) AND a character 
vector with variable names in \code{attr_name}. The variables with the names 
specified in the \code{attr_name} argument will be taken from the  \code{data.frame} 
or \code{matrix} and will be added to the graph as vertex attributes. 
This is useful in case you have a larger \code{data.frame} of which you only 
want to add some of the columns as vertex attributes.
}

In all cases, the attributes will be added for all vertices, so you need to 
make sure the values in the correct order and exist for all of the vertices.
}
\examples{
g <- igraph::graph.ring(5)
add_vertex_attributes(g, "at1", 1:5) |> 
  igraph::get.vertex.attribute("at1")

add_vertex_attributes(g, value = data.frame(naam = 1:5)) |> 
  igraph::get.vertex.attribute("naam")

add_vertex_attributes(g, value = data.frame(een = 11:15, twee = 21:25))

add_vertex_attributes(g, c("twee", "een"), 
  value = data.frame(een = 11:15, twee = 21:25, drie = 31:35))

g <- sna::rgraph(10, tprob = .2) |> network::as.network()
add_vertex_attributes(g, "naam", value = data.frame(naam = 1:10))

add_vertex_attributes(g, value = data.frame(naam = 1:10))

add_vertex_attributes(g, value = data.frame(een = 10:19, twee = 20:29))

add_vertex_attributes(g, c("twee", "een"), 
  value = data.frame(een = 10:19, twee = 20:29, drie = 30:39))
}
