% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_vertex_names}
\alias{add_vertex_names}
\title{Add vertex names to a graph}
\usage{
add_vertex_names(x, value, vids = NULL)
}
\arguments{
\item{x}{graph object to which the vertex attributes need to be added}

\item{value}{the vertex names to be set}

\item{vids}{the id's of the vertices for which the names are to be set.
This is useful if you only want to set/correct the names of a few vertices. 
For \code{igraph} objects, these can be the number or the name of the vertices. 
For \code{network} objects, only the numeric ID is allowed. 
The default is \code{NULL}, which selects all vertices, in the order in which 
they occur in the graph object.}
}
\value{
graph object with new attributes added
}
\description{
Add vertex names to a graph
}
\details{
Consistent API to easily add vertex names to a graph of 
class \code{igraph} or \code{network}. 
The vertex attribute that holds the vertex names is called differently for 
\code{igraph} and \code{network} objects, so 
if you use the incorrect name for the attribute, the new names will not be 
recognized in subsequent steps of the network analysis pipeline. 

This function puts the vertex names in the correct vertex attribute, so you 
do not need to remember what they are called.
}
\examples{
g <- snafun::create_random_graph(5, "gnp", p = .1, graph = "igraph")
# there are no names yet
snafun::extract_vertex_names(g)  # NULL
g <- snafun::add_vertex_names(g, LETTERS[5:1])
snafun::extract_vertex_names(g) # "E" "D" "C" "B" "A"
g <- snafun::add_vertex_names(g, "X", vids = 2)
snafun::extract_vertex_names(g) # "E" "X" "C" "B" "A"
g <- snafun::add_vertex_names(g, c("Y", "Z"), vids = c("C", "A"))
snafun::extract_vertex_names(g) # ""E" "X" "Y" "B" "Z"

g <- snafun::create_random_graph(5, "gnp", p = .1, graph = "network")
snafun::extract_vertex_names(g)  # 1 2 3 4 5
g <- snafun::add_vertex_names(g, LETTERS[5:1])
snafun::extract_vertex_names(g) # "E" "D" "C" "B" "A"
g <- snafun::add_vertex_names(g, "X", vids = 2)
snafun::extract_vertex_names(g) # "E" "X" "C" "B" "A"
# network objects accept only numeric id's
g <- snafun::add_vertex_names(g, c("Y", "Z"), vids = c(3, 5))
snafun::extract_vertex_names(g) # ""E" "X" "Y" "B" "Z"
}
