% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centralityChart.R
\name{centralityChart}
\alias{centralityChart}
\title{centralityChart}
\usage{
centralityChart(
  net,
  measures = c("betweenness", "closeness", "degree"),
  directed = igraph::is.directed(net),
  mode = c("all", "out", "in", "total"),
  normalized = TRUE,
  ...
)
}
\arguments{
\item{net}{a network of class \code{igraph}}

\item{measures}{one or more centrality measures}

\item{directed}{logical, should direction be taken into account? This argument
is optional, if it is not chosen the directionality of the network will be used.}

\item{mode}{Character, used for closeness and degree. The default is "all".}

\item{normalized}{Logical scalar, whether to normalize the centrality scores.
The default is \code{TRUE}, which makes the centralities more motually comparable.}

\item{...}{other arguments passed to \code{\link{graphics}{dotchart}}.}
}
\value{
a plot
}
\description{
Make a chart with centrality scores
}
\details{
This function creates a facet path plot that displays the chosen centrality measures.
Currently, three centrality measures are implemented: "betweenness", "closeness"
, and degree". One or more of these can be chosen.

The \code{mode} is any of "all", "out", "in", "total". Character string,
defines the types of the paths used for measuring the distance in directed
graphs. “in” measures the paths to a vertex, “out” measures paths from a
vertex, all uses undirected paths. This argument is ignored for undirected
graphs.

The names of the vertices are plotted on the y-axis IF a \code{name} attribute
is present for the vertices. Otherwise, their IDs are used.

The default is a point plot. Depending on the neatness of the network object,
it is possible to turn this into a nicer path plot. For this option, set
\code{path} to \code{TRUE}. Depending on the data, this can become a really nice
plot, or become really messy. This is why the default if \code{FALSE}. But feel
free to try it out on a dataset.

The function takes an \code{igraph}-object as input.
}
\examples{
\dontrun{
g <- igraph::erdos.renyi.game(20, 1/20)
centralityChart(g, measures = c("degree", "betweenness"))
# note that closeness is not well-defined for disconnected graphs
centralityChart(g)
}
}
