% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_dyads}
\alias{count_dyads}
\title{Count dyad types}
\usage{
count_dyads(x, echo = TRUE)
}
\arguments{
\item{x}{graph data}

\item{echo}{logical, should the result be printed in the console? Default is 
\code{TRUE}.}
}
\value{
a data.frame
}
\description{
Perform a dyad census
}
\details{
Perform a dyad census on the network data. As input, \code{x} can be 
any of \code{igraph}, \code{network}, or \code{matrix}.

\code{Mutual} represents the number of reciprocated dyads

\code{Asymmetric} represents the number of unreciprocated dyads. This is 
always 0 for undirected graphs.

\code{Null} represents the number of dyads without a relation between the vertices

The output is invisibly returned and (if \code{echo} is \code{TRUE}) 
also printed in the console.
}
\examples{
data(judge_net, package = "snafun")
count_dyads(judge_net)
}
