% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_edges}
\alias{count_edges}
\title{Number of edges in the graph}
\usage{
count_edges(x)
}
\arguments{
\item{x}{graph}
}
\value{
integer
}
\description{
Count the number of edges in the graph
}
\details{
Returns the number of edges in an object of class \code{network} 
or \code{igraph}.
}
\examples{
net <- snafun::create_random_graph(10, "gnp", p = .15, graph = "igraph")
count_edges(net)

net <- snafun::create_random_graph(10, "gnp", p = .15, graph = "network")
count_edges(net)
}
