% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_edges_in_interval.R
\name{count_edges_in_interval}
\alias{count_edges_in_interval}
\title{Count edges in time interval}
\usage{
count_edges_in_interval(x, start = NULL, end = NULL, number = 30)
}
\arguments{
\item{x}{object of class \code{networkDynamic}}

\item{start}{start time for the calculation}

\item{end}{end time for the calculation}

\item{number}{number of intervals in between \code{start} and \code{end}}
}
\value{
vector with counts
}
\description{
Count number of edges in time interval
}
\details{
Counts how many edges happen (ie. have their start within an interval)
over the time period {start, end}. The toime period is split into
\code{number} intervals, each of equal time length. So, if \code{start = 0},
\code{end = 10}, and \code{number = 2}, then the time period is divided
into the time intervals from 0 (inclusive) to 5 (exclusive) and from 5
(inclusive) to 10 (exclusive).

\code{count_edges_in_interval} counts the number of edges that start in
each interval.

\code{count_unique_edges_in_interval} counts the number of unique edges
that start in each interval, so edges that occur multiple times are only
counted once.
}
