% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_triads}
\alias{count_triads}
\title{Count triad types}
\usage{
count_triads(x, echo = TRUE)
}
\arguments{
\item{x}{graph data}

\item{echo}{logical, should the result be printed in the console? Default is 
\code{TRUE}.}
}
\value{
a data.frame
}
\description{
Perform a trad census
}
\details{
Perform a trad census on the network data. As input, \code{x} can be 
any of \code{igraph}, \code{network}, or \code{matrix}.

The output is a data.frame that countains the counts for each of the 16 triad 
types.

The output is invisibly returned and (if \code{echo} is \code{TRUE}) 
also printed in the console.
}
\examples{
data(judge_net, package = "snafun")
count_triads(judge_net)
}
