% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_vertices}
\alias{count_vertices}
\title{Number of vertices in the graph}
\usage{
count_vertices(x)
}
\arguments{
\item{x}{graph}
}
\value{
integer
}
\description{
Count the number of vertices in the graph
}
\details{
Returns the number of vertices in an object of class \code{network} 
or \code{igraph}.
}
\examples{
net <- snafun::create_random_graph(10, "gnp", p = .15, graph = "igraph")
count_vertices(net)

net <- snafun::create_random_graph(10, "gnp", p = .15, graph = "network")
count_vertices(net)
}
