% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_empty_graph}
\alias{create_empty_graph}
\title{Create an empty network}
\usage{
create_empty_graph(n_vertices, directed = TRUE, graph = c("igraph", "network"))
}
\arguments{
\item{n_vertices}{numeric, number of vertices}

\item{directed}{logical, whether to create a directed graph}

\item{graph}{character, the type of graph to be generated: \code{igraph} or 
\code{network}}
}
\value{
a graph of class \code{network} or \code{igraph}
}
\description{
Generate a graph with no edges
}
\details{
Generate a graph with no edges, either as a \code{igraph} or 
\code{network} object.
}
