% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ef_int.R
\name{Ef_int}
\alias{Ef_int}
\title{Ef_int}
\usage{
Ef_int(m, type = "odds")
}
\arguments{
\item{m}{A model object}

\item{type}{Prints either Odds ratios or Probabilities ("odds", "prob")}
}
\value{
a data frame
}
\description{
Estimates the intensity of the effects using Odds ratio and probabilities
}
\details{
This function processes the output of an Exponential Random Graph or a logistic models model and estimates
the intensity of the effects using either Odds ratios or Probabilities
The function returns a data frame with four colums:
- Estimates
- Either Odds ratios or Probabilities
- Standard Errors
- P-values
}
\examples{
\dontrun{
flo <-SNA4DSData::florentine
fflom <- flo$flomarriage
flom <- to_network(fflom)
m <- ergm::ergm(flom ~ edges + nodecov("Wealth"))

Ef_int(m, prob)
}
}
