% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_all_vertex_attributes.R
\name{extract_all_vertex_attributes}
\alias{extract_all_vertex_attributes}
\title{Extract all vertex attributes}
\usage{
extract_all_vertex_attributes(g)
}
\arguments{
\item{g}{name of the input network}
}
\value{
data.frame in case of vertex attributes, otherwise NULL
}
\description{
Extract all vertex attributes from the network
}
\details{
This function extract all vertex attributes from a \code{network} object
or from a \code{igraph} object and returns it as a data.frame.

The columns of the data.frame have the corresponding attribute names.

No attempt is made to assign a specific class to the attributes. The user
will have to set them appropriately for the analysis at hand.

Note that \code{network} objects tend to have a vertex attribute called
\code{na}. This will, correctly, be included, but is likely irrelevant.

If there are no vertex attributes, \code{NULL} is returned.
}
\examples{
\dontrun{
data(florentine, package = "SNA4DSData")
flomar <- florentine$flomarriage
extract_all_vertex_attributes(flomar)

# NULL
extract_all_vertex_attributes(igraph::erdos.renyi.game(10, 4, type = "gnm"))

flomar_nw <- intergraph::asNetwork(flomar)
extract_all_vertex_attributes(flomar_nw)
}
}
