% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_cut_vertices}
\alias{extract_cut_vertices}
\title{Extract cut vertices}
\usage{
extract_cut_vertices(x, names = TRUE)
}
\arguments{
\item{x}{graph object of class \code{igraph}, \code{network}, \code{matrix},
or \code{data.frame}}

\item{names}{logical, should symbolic vertex names be returned when
available? Defaults to \code{TRUE}.}
}
\value{
vector of vertex ids or names
}
\description{
Identify the articulation points of a graph.
}
\details{
A cut vertex is a vertex whose removal increases the number of connected
components. For directed graphs, edge directions are ignored, matching the
usual articulation-point interpretation.
}
\examples{
g <- snafun::create_manual_graph(A -- B -- C -- D)
extract_cut_vertices(g)
extract_cut_vertices(snafun::to_matrix(g))
}
