% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_loops}
\alias{extract_loops}
\alias{extract_loops_vertex}
\title{Extract loops}
\usage{
extract_loops(x)

extract_loops_vertex(x)
}
\arguments{
\item{x}{input graph of class \code{igraph} or \code{network}}
}
\value{
see above
}
\description{
Extract the edges that run from an actor to himself
}
\details{
Several functions in this package deal with loops (also sometimes redundantly 
called "self-loops")"ties that run from a vertex to that same vertex. 

This specific function extracts the edge id's for these loops, so they can be 
identified and, if required, removed.

\code{extract_loops} returns the edge ID's of the loops in the graph

\code{extract_loops_vertex} returns a table with two columns, the first 
contains the vertex that has at least one loop in the graph and the 
second gives the number of loops for that vertex in the graph.
}
\examples{
data(florentine, package = "snafun")
x <- florentine$flobusiness
has_loops(x)             # FALSE
extract_loops(x)         # NULL
extract_loops_vertex(x)  # NULL
x <- igraph::add_edges(x, c("Barbadori", "Barbadori", "Medici", "Medici"))
has_loops(x)             # FALSE
extract_loops(x)         # loops detected
extract_loops_vertex(x)
}
