% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_cug_input.R
\name{fix_cug_input}
\alias{fix_cug_input}
\title{Fix intermediate CUG input}
\usage{
fix_cug_input(el, directed = TRUE)
}
\arguments{
\item{el}{the input generated by a single run of \code{sna::cug.test}}

\item{directed}{logical, is the graph directed or not?}
}
\value{
object of class \code{igraph}
}
\description{
Fix the intermediate input from sna::cug.test
}
\details{
This function makes the \link[sna]{cug.test} useable for \code{snafun}. 
The cug.test function generates a series of networks in a specific way. 
This function converts that data format into an object of class \code{igraph}, 
which can then be handled by (almost) any function in \code{snafun} or 
\code{igraph} or other code that works with \code{igraph} objects.

This does slow down the computations in cug.test a little, but the added 
flexibility is a huge gain.

Make sure to set correctly whether the actual graph is directed or not.
}
