% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{florentine}
\alias{florentine}
\title{Florentine families}
\format{
A list containing two \code{igraph} networks and a data.frame with
attributes.
}
\source{
http://networkdata.ics.uci.edu/netdata/html/florentine.html
(plain text on purpose: this stale external link is kept for reference but
not marked up as a link, so R CMD check does not fetch it)
}
\usage{
data(florentine, package = "snafun")
}
\description{
Network data of 16 Florentine families, in \code{igraph} format.
}
\details{
A list containing two networks and a data.frame with attributes.

This is a data set of 1520 marriage and 15 business ties among Renaissance Florentine
families. The data is originally from Padgett (1994) via UCINET and stored
as an \code{igraph} object.

This dataset is a subset of the original data and contains data on sixteen
families.

The two relations are business ties (\code{flobusiness}, recorded financial
ties such as loans, credits and joint partnerships) and
marriage alliances (\code{flomarriage}).

In addition, there is a data.frame with attributes:
\itemize{
\item wealth each family's net wealth in 1427 (in thousands of lira)
\item priorates the number of priorates (seats on the civic council) held
between 1282- 1344
\item totalties the total number of business or marriage ties in the total
dataset of 116 families.
}

The data are symmetrically coded. This makes sense for marital ties,
but is unfortunate for the financial ties (which are almost certainly
directed in real life).

Substantively, the data include families who were engaged in a struggle
for political control of the city of Florence in around 1430.
Two factions were dominant in this struggle: one revolved around the
infamous Medicis (9), the other around the powerful Strozzis (15).
}
\keyword{internal}
