% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{GOF_plot}
\alias{GOF_plot}
\title{GOF plot}
\usage{
GOF_plot(gof)
}
\arguments{
\item{gof}{A gof model object for a fitted \code{ergm} or \code{btergm} model}
}
\value{
Plots displaying the Goodness of Fit
}
\description{
Plots the Goodness of fit for the output of a model of class \code{ergm} or \code{btergm}
}
\details{
This function plots the goodness-of-fit of unipartite Exponential Random Graphs Models 
or BTERGM models.

It does not process bipartite ERGMs model objects nor other types of 
Exponential random graph models object.
}
\examples{
\dontrun{
### ergm model
flo <-SNA4DSData::florentine
flom <- intergraph::asNetwork(flo$flomarriage)
m_ergm <- ergm::ergm(flom ~ edges + nodecov("Wealth"))
gof_ergm <- ergm::gof(m_ergm)
GOF_plot(gof_ergm)

### btergm model
data(alliances, package = "SNA4DSData")
m_btergm <- btergm::btergm(allianceNets ~ edges + 
  gwesp(0, fixed = TRUE) +
  edgecov(lastYearsSharedPartners) + 
  edgecov(militaryDisputes) + 
  nodecov("polity") +
  nodecov("Composite_Index_National_Capability") + 
  absdiff("polity") + 
  absdiff("Composite_Index_National_Capability") +
  edgecov(sharedBorder), 
  R = 100, # number of bootstraps
  parallel = "snow", ncpus = 16  # optional line
)

gof_btergm <- gof_btergm(m_btergm)
GOF_plot(gof_btergm)
}
}
