% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R
\name{has_edge_attributes}
\alias{has_edge_attributes}
\title{Check for the existence of edge attributes in the graph}
\usage{
has_edge_attributes(x)
}
\arguments{
\item{x}{graph of class \code{igraph} or \code{network}}
}
\value{
logical
}
\description{
Check for the existence of edge attributes in the graph
}
\details{
Simple utility function that returns \code{TRUE} if the graph has 
edge attributes and \code{FALSE} otherwise.
}
