% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R
\name{has_vertexnames}
\alias{has_vertexnames}
\title{has_vertexnames}
\usage{
has_vertexnames(x)
}
\arguments{
\item{x}{the network, either of class \code{igraph} or \code{network}}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Check whether the object contais vertex names
}
\details{
A simple check, yielding \code{TRUE} or \code{FALSE}. It merely checks 
for the presence of the vertex attribute \code{name} (for a network of
class \code{igraph}) or \code{vertex.names} (for a network of
class \code{network}). 

Note that the \code{network} package tends to create vertex names by 
default, even if they were not added as separate attributes. 
In that case, the vertices are named as integers running  from 1 to the 
number of vertices. Of course, this function can not distinguish 
whether these are names the researcher want to use as names, or purely 
the result of the default behavior of the \code{network} package (and 
not meaningful names per se).
}
\examples{
data(florentine, package = "snafun")
has_vertexnames(florentine$flobusiness)
}
