% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_directed}
\alias{is_directed}
\title{Is the network directed?}
\usage{
is_directed(x)
}
\arguments{
\item{x}{graph x}
}
\value{
logical, \code{TRUE} or \code{FALSE}
}
\description{
Check if the network is directed
}
\details{
This function checks if \code{x} is a directed network.

For an x of class \code{igraph} or \code{network}, the function checks 
for the appropriate attribute in the graph x.

For a \code{matrix} the function returns \code{TRUE} if the matrix is 
not symmetric.
 
For a \code{data.frame} the function returns \code{TRUE} if the reciprocity 
of the network is not exactly 0 or 1.
}
\examples{
data(florentine, package = "snafun")
is_weighted(florentine$flobusiness)   # FALSE
}
\references{
largely based on similar functions from the excellent 
\code{migraph} package. 
These methods are licensed through the 
\href{https://choosealicense.com/licenses/mit/}{MIT license}
in combination with copyright for James Hollway.
}
