% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_weighted}
\alias{is_weighted}
\title{Is the network weighted?}
\usage{
is_weighted(x)
}
\arguments{
\item{x}{graph x of class \code{network}, \code{igraph}, 
\code{matrix}, or \code{data.frame}}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Check if the network is weighted
}
\details{
This function checks if \code{x} is a weighted network.

For an x of class \code{igraph}or \code{network}, the function checks 
for the appropiate attribute in the graph x.

For a \code{matrix} the function returns \code{TRUE} if any of the 
elements is not or 1.
 
For a \code{data.frame} the function returns \code{TRUE} if it contains 
a numeric column called "weight", in addition to the first two columns 
of senders and receivers.
}
