% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Madrid_bombing}
\alias{Madrid_bombing}
\title{Madrid Train Bombing 2004 (dataset)}
\format{
list of igraph objects
}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks) 
or here http://doitapps.jjay.cuny.edu/jjatt/data.php. Converted to R format 
in the \code{networkdata} package from David Schoch.
}
\usage{
data(Madrid_bombing, package = "snafun")
}
\description{
This is a time series that treat specific attacks as endpoints 
and depict the evolution of relations between individuals indirectly and directly 
associated with the Madrid train bombing. http://en.wikipedia.org/wiki/2004_Madrid_train_bombings

1-mode stacked matrices 55 x 55 person by person, data on 20 time periods plus kinship data and tie extinguished data.

Codebook available here http://doitapps.jjay.cuny.edu/jjatt/files/Relations_Codebook_Public_Version2.pdf
}
\keyword{internal}
