% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_snafun_cheatsheet.R
\name{open_SNA4DS_cheatsheet}
\alias{open_SNA4DS_cheatsheet}
\title{Open a SNA4DS cheatsheet in the browser}
\usage{
open_SNA4DS_cheatsheet(graphics = TRUE)
}
\arguments{
\item{graphics}{logical, should the list of options be shown as a clickable
graphical menu?}
}
\value{
nothing
}
\description{
Open a SNA4DS cheatsheer in your browser
}
\details{
Shows a list of currently available online vignettes for the 
\code{SNA4DS} course.

The user can pick the preferred vignette by entering the number that corresponds
to the preferred vignette. The vignette will then open in the user's default
web browser.

When \code{graphics} is \code{TRUE}, a graphical choice menu is shown. If that
is not preferred, or if the user's machine lacks the graphical tools needed,
setting \code{graphics} to \code{FALSE} will show the list of vignettes in the
R console.
}
\examples{
\dontrun{
open_SNA4DS_cheatsheet()
}
}
