% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_qap_lm.R
\name{plot.stat_qap_lm}
\alias{plot.stat_qap_lm}
\title{Plot a QAP linear model result}
\usage{
\method{plot}{stat_qap_lm}(
  x,
  term = 1,
  statistic = c("beta", "t-value"),
  main = NULL,
  sub = NULL,
  breaks = "Sturges",
  col = "grey85",
  border = "white",
  observed_col = "firebrick",
  observed_lwd = 2,
  xlab = NULL,
  prob = TRUE,
  ...
)
}
\arguments{
\item{x}{Object returned by \code{\link{stat_qap_lm}}.}

\item{term}{Coefficient to plot. Either a numeric index or a coefficient
name.}

\item{statistic}{Which replicate distribution should be plotted? One of
\code{"beta"} or \code{"t-value"}.}

\item{main}{Plot title.}

\item{sub}{Plot subtitle.}

\item{breaks}{Histogram break specification.}

\item{col}{Fill color for the histogram bars.}

\item{border}{Border color for the histogram bars.}

\item{observed_col}{Color of the observed-value reference line.}

\item{observed_lwd}{Line width of the observed-value reference line.}

\item{xlab}{X-axis label.}

\item{prob}{Logical scalar; should a density histogram be drawn?}

\item{...}{Additional graphical arguments passed to
\code{\link[graphics]{hist}}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Plot the permutation distribution for a selected coefficient from
\code{\link{stat_qap_lm}}.
}
