% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comm_dendrogram.R
\name{plot_comm_dendrogram}
\alias{plot_comm_dendrogram}
\title{Plot a dendrogram of the community structure}
\usage{
plot_comm_dendrogram(
  x,
  labels = NULL,
  hang = 0.1,
  axes = TRUE,
  frame_plot = FALSE,
  ann = TRUE,
  main = "Communities dendrogram",
  sub = "",
  xlab = "",
  ylab = "",
  ...
)
}
\arguments{
\item{x}{a \code{communities} object}

\item{labels}{A character vector of labels for the leaves of the tree. 
By default the row names or row numbers of the original data are used. 
If \code{labels = FALSE} no labels at all are plotted.}

\item{hang}{The fraction of the plot height by which labels should hang 
below the rest of the plot. A negative value will cause the labels to hang 
down from 0.}

\item{axes, frame_plot, ann}{logical flags as in \code{\link[graphics]{plot.default}}.}

\item{main, sub, xlab, ylab}{character strings for \code{\link[graphics]{title}}.}

\item{...}{Further graphical arguments. E.g., \code{cex} controls the size of 
the labels (if plotted) in the same way as \code{\link[graphics]{text}}.}
}
\value{
nothing, only the plot is shown
}
\description{
Plot a hierarchical cluster analysis on a communities object
}
\details{
Plot the result of one of the community detection algorithms as a nice 
dendrogram. Except for \code{x}, all arguments are used for the \code{plot} 
function.

This works for the results from the \code{\link{extract_comm_walktrap}}, 
\code{\link{extract_comm_fastgreedy}}, and \code{\link{extract_comm_girvan}}. 
It (currently) does not work for the results from \code{\link{extract_comm_louvain}}.
}
