% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_nam.R
\name{plot_nam}
\alias{plot_nam}
\title{Plot diagnostics for the network autorrelation model}
\usage{
plot_nam(x, ...)
}
\arguments{
\item{x}{output from \code{\link[snafun]{stat_nam}}}

\item{...}{ignored}
}
\value{
nothing is returned, the plots are displayed in the console
}
\description{
Plot diagnostics for the network autorrelation model
}
\details{
Plots diagnostics For the output of \code{\link[snafun]{stat_nam}}.
}
