% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network_slices.R
\name{plot_network_slices}
\alias{plot_network_slices}
\title{Plot network slices}
\usage{
plot_network_slices(x, number = 9, start = NULL, end = NULL, digits = 3)
}
\arguments{
\item{x}{object of class \code{networkDynamic}}

\item{number}{number of intervals in between \code{start} and \code{end}}

\item{start}{start time for the calculation}

\item{end}{end time for the calculation}

\item{digits}{number of decimals, used to show the time slices in the plot}
}
\value{
plot
}
\description{
Plot slices of a \code{networkDynamic} network
}
\details{
This function is a variation of \link[ndtv]{filmstrip}.
That function has the odd behavior of only plotting edges that occur
at specific times, rather than over intervals.
The current function plots the edges that occur within time intervals.

This function splits the time interval between the starting time
(\code{start}) and the end time (\code{end}),
into \code{number} equal-sized intervals.
The plots include all edges that occur in that interval.

By default, the overall interval covers the period from where the first
edges starts to where the last one ends.

Each interval is left-closed and right-open.
This means that it can happen that the very last edge is not included, if
it starts and ends at the very last time point.
This can be solved by setting the \code{end} time (a little) higher than
the actual end time.
}
