% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_helpers.R
\name{print.communities}
\alias{print.communities}
\title{Print a community object}
\usage{
\method{print}{communities}(x, digits = 4, ...)
}
\arguments{
\item{x}{object of class \code{communities}}

\item{digits}{number of significant digits to print}

\item{...}{ignored}
}
\value{
invisibly returns \code{x}
}
\description{
Print the main characteristics of a \code{communities} object in a compact
student-friendly format.
}
\examples{
data("judge_net", package = "snafun")
coms <- snafun::extract_comm_fastgreedy(judge_net)
print(coms)
}
