% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_helpers.R
\name{print.summary.communities}
\alias{print.summary.communities}
\title{Print a community summary}
\usage{
\method{print}{summary.communities}(x, digits = 4, ...)
}
\arguments{
\item{x}{object of class \code{summary.communities}}

\item{digits}{number of significant digits to print}

\item{...}{ignored}
}
\value{
invisibly returns \code{x}
}
\description{
Print the object returned by \code{\link[base:summary]{summary()}} for a
\code{communities} object.
}
\examples{
data("judge_net", package = "snafun")
coms <- snafun::extract_comm_fastgreedy(judge_net)
print(summary(coms))
}
