% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_cug.R
\name{print.summary.stat_cug}
\alias{print.summary.stat_cug}
\title{Print a summary of a CUG test result}
\usage{
\method{print}{summary.stat_cug}(x, digits = 4, ...)
}
\arguments{
\item{x}{Object returned by \code{\link{summary.stat_cug}}.}

\item{digits}{Number of significant digits to print.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print the output created by \code{\link{summary.stat_cug}}.
}
