% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_qap_lm.R
\name{print.summary.stat_qap_lm}
\alias{print.summary.stat_qap_lm}
\title{Print a summarized QAP linear model}
\usage{
\method{print}{summary.stat_qap_lm}(x, digits = 4, ...)
}
\arguments{
\item{x}{Object returned by \code{\link{summary.stat_qap_lm}}.}

\item{digits}{Number of significant digits to print.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print the object returned by \code{\link{summary.stat_qap_lm}}.
}
