% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_qap_logit.R
\name{print.summary.stat_qap_logit}
\alias{print.summary.stat_qap_logit}
\title{Print a summarized QAP logistic network model}
\usage{
\method{print}{summary.stat_qap_logit}(x, digits = 4, ...)
}
\arguments{
\item{x}{Object returned by \code{\link{summary.stat_qap_logit}}.}

\item{digits}{Number of significant digits to print.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print the object returned by \code{\link{summary.stat_qap_logit}}.
}
