% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove}
\alias{remove}
\alias{remove_edge_attribute}
\alias{remove_vertex_attribute}
\alias{remove_graph_attribute}
\title{Remove parts of the graph}
\usage{
remove_edge_attribute(x, attr_name)

remove_vertex_attribute(x, attr_name)

remove_graph_attribute(x, attr_name)
}
\arguments{
\item{x}{input graph (class \code{igraph} or \code{network})}

\item{attr_name}{name of the attribute to be removed}
}
\value{
graph, without the removed elements
}
\description{
Remove parts of the graph
}
\examples{
data(emon, package = "network")
texas <- emon$Texas
remove_edge_attribute(texas, "Frequency")
remove_vertex_attribute(texas, "Formalization")
remove_graph_attribute(texas, "directed")

texas_i <- to_igraph(texas)
remove_edge_attribute(texas_i, "Frequency")
remove_vertex_attribute(texas_i, "Formalization")
}
