% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_isolates}
\alias{remove_isolates}
\title{Remove all isolates from the graph}
\usage{
remove_isolates(x, loops = FALSE)
}
\arguments{
\item{x}{graph of class \code{igraph} or \code{network}}

\item{loops}{should self-loops count when deciding if a vertex is an isolate? 
See  \code{\link{find_isolates}} for details.}
}
\value{
graph with the isolates (if any) removed
}
\description{
Remove all isolates from the graph
}
\details{
Removes the isolates from an object of class \code{igraph} or \code{network}.
First, the algorithm identifies the isolates, using the \code{\link{find_isolates}} 
function. Then, the identified isolates are removed and the new, smaller, 
graph is returned. Obviously, if \code{x} does not contain any isolates, 
\code{x} will be returned unaltered.
}
\seealso{
Finding isolates: \code{\link{find_isolates}}
}
