% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_loops}
\alias{remove_loops}
\title{Remove loops}
\usage{
remove_loops(x)
}
\arguments{
\item{x}{graph of class \code{igraph}, \code{network}, or \code{matrix}}
}
\value{
graph with the same class as the input
}
\description{
Remove loops from the graph
}
\details{
Remove loops (ie. a tie from a vertex to itself) from the graph object
}
\examples{
x <- matrix(c(1, 1, 0, 0, 0, 1, 1, 0, 1), ncol = 3, byrow = TRUE)
remove_loops(x)
g_n <- snafun::to_network(x)
snafun::to_matrix(g_n)
remove_loops(g_n) |> snafun::to_matrix()
remove_loops(snafun::to_igraph(x)) |> snafun::to_matrix()
}
