% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{soccer98}
\alias{soccer98}
\title{Soccer98 network}
\format{
Object of class network of size 36, directed, weighted, named.
}
\source{
Data collected by Lothar Krempel, October 5, 1999.
Transformed in Pajek format by V. Batagelj, February 9, 2001.
Raw data are available at the Pajek datasets (http://vlado.fmf.uni-lj.si/pub/networks/data/sport/football.htm).
}
\usage{
data(soccer98, package = "snafun")
}
\description{
The network is based on the 32 soccer teams that participated in the 
World Championship soccer in Paris, 1998. 
Players of the national team often have contracts in other countries. 
This constitutes a players market where national teams "export" players to other 
countries. Members of the 32 teams had contracts in 36 countries.
}
\details{
The vertices are of the network are the 36 countries that players of these 
national teams play in. The senders are the countries that export 
players, the receivers are the countries that import players.

The edge weights *i, j* represent the number of players of nationality *i* 
that play in country *j* at the time of the World Championship.

There exist countries that are only exporters: Yugoslavia, Chile, Cameroon, Nigeria, ...
Some other countries are only importers: Spain, France, Turkey, Greece, GBR...
}
\keyword{internal}
