% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_nam.R
\name{stat_nam_summary}
\alias{stat_nam_summary}
\title{Network autocorrelation results}
\usage{
stat_nam_summary(
  x,
  correlation = TRUE,
  R2 = TRUE,
  digits = 3,
  signif.stars = TRUE
)
}
\arguments{
\item{x}{the output object from \code{\link[snafun]{stat_nam}}}

\item{correlation}{logical; if \code{TRUE} (the default), the correlation 
matrix of the estimated parameters including sigma is returned and printed}

\item{R2}{logical; if \code{TRUE} (the default), the Nagelkerke 
pseudo R-squared is reported}

\item{digits}{the number of significant digits to use when printing}

\item{signif.stars}{;ogical; if \code{TRUE} (the default), "significance stars" 
are printed for each coefficient}
}
\value{
prints the summary to the console and invisibly returns the 
summary object
}
\description{
Extract the summary of a network autocorrelation model analysis
}
\details{
This function wraps functions from the \pkg{spatialreg} package to summarize 
the results from an analysis with the \code{\link[snafun]{stat_nam}} function.
}
