% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_equivalence.R
\name{structural_equivalence}
\alias{structural_equivalence}
\title{Structural equivalence}
\usage{
structural_equivalence(g, weight = NULL, digits = 3, suppressWarnings = TRUE)
}
\arguments{
\item{g}{graph of class \code{igraph}}

\item{weight}{Either \code{NULL} (the default) or a character string giving 
an edge attribute name. 
If the graph has multiple edges, the edge attribute of an arbitrarily chosen 
edge (for the multiple edges) is included. 
Note that the function requires the attribute to be either logical or numeric.}

\item{digits}{number of decimals to use}

\item{suppressWarnings}{logical, whether or not any warnings should be returned}
}
\value{
matrix of size n*n, with n equal to the number of vertices in the graph
}
\description{
Calculate structural equivalence
}
\details{
Calculate structural equivalence based on correlations. 
The function, of course, disregards the diagonal.

Vertices are also perfectly structurally equivalent with 
themselves.

If requested, an edge attribute can be taken into account 
through the \code{weight} argument. By default, no weights 
are used.

The value varies between \[-1, 1\], with -1 denoting maximally 
possible distance between two vertices and +1 denoting 
exact structural equivalence.
}
\examples{
\dontrun{
#' data("judge_net", package = "SNA4DSData")
structural_equivalence(judge_net)
}
}
