% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_qap_logit.R
\name{summary.stat_qap_logit}
\alias{summary.stat_qap_logit}
\title{Summarize a QAP logistic network model}
\usage{
\method{summary}{stat_qap_logit}(object, ...)
}
\arguments{
\item{object}{Object returned by \code{\link{stat_qap_logit}}.}

\item{...}{Ignored.}
}
\value{
An object of class \code{summary.stat_qap_logit}.
}
\description{
Summarize the coefficient-level results from \code{\link{stat_qap_logit}}.
}
