% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_binary_matrix.r
\name{to_binary_matrix}
\alias{to_binary_matrix}
\title{Dichotomize a matrix}
\usage{
to_binary_matrix(mat, min)
}
\arguments{
\item{mat}{a numeric matrix}

\item{min}{a threshold. All values in \code{mat} of at least this minimum value 
will become 1, all values smaller than this minimum will become 0.}
}
\value{
the dichotomized matrix
}
\description{
Make a binary matrix out of a weighted matrix
}
\details{
If a cell has value smaller than a minimum value, 
the cell becomes 0. If the cell has a value of at least 
the minimum value, it becomes 1.
The procedure will work for any matrix, not necessarily a square one.
}
\examples{
M <- matrix(round(runif(20),1), ncol = 4)
to_binary_matrix(M, min = .45)
}
