% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_edgelist.R
\name{to_edgelist}
\alias{to_edgelist}
\title{Make edgelist}
\usage{
to_edgelist(x, named = TRUE, sort = "from")
}
\arguments{
\item{x}{input object}

\item{named}{logical, should vertex names be returned (if they exist) (\code{TRUE})
or not (\code{FALSE})}

\item{sort}{character, name of the column to sort on. Defaults to "from".}
}
\value{
data frame
}
\description{
Make an edgelist from a set of input objects. This function doesn't do 
anything that \code{igraph} or \code{network}/\code{sna} can't also already 
do. The usefulness of this function is its consistent API.
}
\details{
Currently, input can be a matrix, \code{igraph} object or \code{network} object.

When the input is a graph object, the returned data.frame silently stores
vertex metadata as attributes. This allows \code{snafun::to_igraph()},
\code{snafun::to_network()}, and \code{snafun::to_matrix()} to reconstruct
isolates during a roundtrip through an edgelist. These hidden attributes are
intentionally lightweight and will often disappear when the edgelist is
heavily edited, subsetted, or written to disk, in which case the conversion
falls back to the information visible in the edgelist itself.
}
\examples{
## from an igraph object
g_i <- snafun::create_random_graph(10, "gnp", p = 2/10, graph = "igraph")
to_edgelist(g_i)

# add vertex names
g1_i <- g_i
g1_i <- snafun::add_vertex_names(
  g1_i,
  LETTERS[seq_len(snafun::count_vertices(g1_i))]
)
to_edgelist(g1_i)
# add weights
g1_i <- snafun::add_edge_attributes(
  g1_i,
  "weight",
  value = seq_len(snafun::count_edges(g1_i))
)
to_edgelist(g1_i)
# add further edge attribute
g2_i <- snafun::add_edge_attributes(
  g1_i,
  "color",
  value = rep("red", snafun::count_edges(g1_i))
)
to_edgelist(g2_i)

## from a matrix
g_n <- snafun::to_matrix(
  snafun::create_random_graph(10, "gnp", p = .2, graph = "igraph")
)
to_edgelist(g_n)
g_b <- snafun::create_bipartite(5, 3, strategy = "gnm", m = 6)
to_edgelist(g_b)

## from a network object
g2_n <- snafun::to_network(g_n)
to_edgelist(g2_n)
}
