% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_geokpath_w.R
\name{v_geokpath_w}
\alias{v_geokpath_w}
\title{Geodesic k-path centrality}
\usage{
v_geokpath_w(
  graph,
  vids = NULL,
  mode = c("all", "out", "in"),
  weights = NULL,
  k = 3
)
}
\arguments{
\item{graph}{The input graph as igraph object}

\item{vids}{Numeric vertex sequence, the vertices that should be considered.
Default is all vertices. Otherwise, the operation is performed on the 
subgraph only containing vertices \code{vids}.}

\item{mode}{Character constant, gives whether the shortest paths to or from 
the given vertices should be calculated for directed graphs. 
If \code{out} then the shortest paths from the vertex, if \code{in} then to 
it will be considered. If \code{all}, the default, then the corresponding 
undirected graph will be used, ie. not directed paths are searched. 
This argument is ignored for undirected graphs.}

\item{weights}{Possibly a numeric vector giving edge weights. 
If this is \code{NULL}, the default, and the graph has a weight edge attribute, 
then the attribute is used. If this is \code{NA} then no weights are used 
(even if the graph has a weight attribute).}

\item{k}{The k parameter. The default is 3.}
}
\value{
A numeric vector contaning the centrality scores for the selected vertices.
}
\description{
Geodesic K-path centrality counts the number of vertices that can be reached 
through a geodesic path of length less than "k".
}
\details{
This function counts the number of vertices that a specific vertex can 
reach within k steps. 
By default, this number is weighted (if the graph has a \code{weight} 
edge attribute). 
This can be overridden by setting the \code{weights} argument to 
\code{NA} (no weight is used) or to a vector with weights (typically 
this is a numeric edge attribute).

More detail at Geodesic K-Path Centrality
(http://www.centiserver.org/?q1=centrality&q2=Geodesic_K-Path_Centrality)
}
\examples{
g <- igraph::barabasi.game(100)
v_geokpath_w(g)
}
\references{
Borgatti, Stephen P., and Martin G. Everett. 
"A graph-theoretic perspective on centrality." Social networks 28.4 (2006): 466-484.
}
\author{
Mahdi Jalili \email{m_jalili@farabi.tums.ac.ir} (adapted for this package)
}
