% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\alias{extract_vertex_attribute}
\alias{extract_vertex_names}
\alias{extract_edge_attribute}
\alias{extract_graph_attribute}
\title{Extract attributes from the graph object}
\usage{
extract_vertex_attribute(x, name)

extract_vertex_names(x)

extract_edge_attribute(x, name)

extract_graph_attribute(x, name)
}
\arguments{
\item{x}{graph object of class \code{igraph} or \code{network}}

\item{name}{name of the attribute to be accessed}
}
\value{
the values of the requested attributes (if any)
}
\description{
Extract attributes from the graph object
}
