% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_bipartite}
\alias{is_bipartite}
\title{Is the network bipartite?}
\usage{
is_bipartite(x)
}
\arguments{
\item{x}{graph x of class \code{network} or \code{igraph}}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Check if the network is bipartite
}
\details{
This function simply checks if \code{x} is a bipartite network.
This is only possible by checking whether the bipartite attributes are 
set in the x. 
Hence, if the network is bipartite, but the researcher neglected to include 
this information in the graph x, the result of this functions will be 
\code{FALSE}. Of course, this is how it should be: neglecting to include 
the bipartite attribute in the x is bad practice.
}
\examples{
net <- igraph::sample_bipartite(10, 5, p =.1)
is_bipartite(net)  # TRUE

net <- igraph::erdos.renyi.game(10, p.or.m = .1, type = "gnp")
is_bipartite(net)  # FALSE

mat <- sna::rgraph(10, m = 1, tprob = .1)
net <- network::as.network.matrix(mat)
is_bipartite(net)  # FALSE
}
