% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_attributes}
\alias{list_attributes}
\alias{list_vertex_attributes}
\alias{list_edge_attributes}
\alias{list_graph_attributes}
\title{List the attributes in the graph object}
\usage{
list_vertex_attributes(x)

list_edge_attributes(x)

list_graph_attributes(x)
}
\arguments{
\item{x}{graph object of class \code{igraph} or \code{network}}
}
\value{
the names of the requested attributes (if any)
}
\description{
List the attributes in the graph object using a consistent API
}
