% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.igraph}
\alias{plot.network}
\title{Plot the graph object}
\usage{
\method{plot}{igraph}(x, ...)

\method{plot}{network}(x, ...)
}
\arguments{
\item{x}{graph object of class \code{network} or \code{igraph}}

\item{...}{Additional arguments.}
}
\value{
nothing, the object is merely plotted
}
\description{
Plot the graph object
}
\details{
Plots a graph object of class \code{network} or \code{igraph} exactly as it 
would be plotted by the \code{network} or \code{igraph} packages, 
respectively. The function wraps \link[network]{plot.network} and 
\link[igraph]{plot.igraph}.  

Whatever is set in \code{...} is passed onto these original functions. 
See the respective help pages (\link[network]{plot.network}, 
\link[igraph]{plot.igraph}, and \link[igraph]{igraph.plotting}) 
if you want to use settings that deviate from 
the default settings in these packages (which is almost always better than 
using the defaults).

This function plots the objects just as they would be plotted if the 
packages were attached, but now they do not need to be attached.
}
\examples{
g_i <- snafun::create_random_graph(10, "gnm", m = 20, graph = "igraph")
g_n <- snafun::create_random_graph(10, "gnm", m = 20, graph = "network")
plot(g_i)
plot(g_i, vertex.size = 12, vertex.color = "green", edge.width = 5, edge.curved = TRUE)
plot(g_n)
plot(g_n, vertex.cex = 3, vertex.col = "green", edge.lwd = 10, 
   edge.col = "darkgrey", usecurve = TRUE, edge.curve = .05, 
   arrowhead.cex = 3, displaylabels = TRUE, label.pos = 5)
}
