% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.igraph}
\alias{print.network}
\title{Print the graph object}
\usage{
\method{print}{igraph}(x, ...)

\method{print}{network}(x, na.omit = FALSE, ...)
}
\arguments{
\item{x}{graph object of class \code{network} or \code{igraph}}

\item{...}{Additional agruments.}

\item{na.omit}{logical; omit summarization of missing attributes in a 
\code{network} object?}
}
\value{
nothing, the object is merely printed
}
\description{
Print the graph object
}
\details{
Prints a graph object of class \code{network} or \code{igraph} exactly as it 
would be printed by the \code{network} or \code{igraph} packages, 
respectively. The function wraps \link[network]{print.network} and 
\link[igraph]{print.igraph}.  
Whatever is set in \code{...} is passed onto these original functions. 
See the respective help pages if you want to use settings that deviate from 
the default settings in these packages.

This function prints the objects just as they would be printed if the 
packages were attached, but now they do not need to be attached.
}
\section{Methods (by class)}{
\itemize{
\item \code{print(igraph)}: S3 method to print an igraph object

\item \code{print(network)}: S3 method to print an igraph object

}}
\examples{
g_i <- snafun::create_random_graph(10, "gnm", m = 20, graph = "igraph")
g_n <- snafun::create_random_graph(10, "gnm", m = 20, graph = "network")
print(g_i)
print(g_n)
}
