% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snafun-package.r
\docType{package}
\name{snafun}
\alias{snafun}
\alias{snafun-package}
\title{snafun}
\description{
Package supporting the SNA4DS course
}
\details{
Functions, useful for the SNA4DS course. A series of functions that provide 
a (fairly) consistent API to deal with \code{network} and \code{igraph} 
graph objects. It reduces the need (well, some of it) to constantly switch 
between the \code{network}, \code{sna}, and \code{igraph} packages, as is 
typical of most real life network analysis projects.
Where that switch is still needed, the \code{snafun} makes it much less 
painful and, yes, even fun.
In short: SNA FUN!
}
\keyword{package}
