% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_ef_int.R
\name{stat_ef_int}
\alias{stat_ef_int}
\title{Estimate the intensity of effects using odds ratio and probabilities}
\usage{
stat_ef_int(m, type = "odds")
}
\arguments{
\item{m}{A model object}

\item{type}{Prints either Odds ratios or Probabilities ("odds", "prob")}
}
\value{
a \code{data.frame}
}
\description{
Estimates the intensity of the effects using odds ratio and probabilities
}
\details{
This function processes the output of an Exponential Random Graph or a logistic models model and estimates
the intensity of the effects using either Odds ratios or Probabilities
The function returns a data frame with four colums:
- Estimates
- Either Odds ratios or Probabilities
- Standard Errors
- P-values
}
\examples{
\dontrun{
data(florentine, package = "snafun")
fflom <- florentine$flomarriage
flom <- to_network(fflom)
m <- ergm::ergm(flom ~ edges + nodecov("Wealth"))

stat_ef_int(m, "prob")
}
}
\seealso{
Other statistics functions: 
\code{\link{stat_plot_gof}()},
\code{\link{stat_plot_gof_as_btergm}()}
}
\concept{statistics functions}
