% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_gof.R
\name{stat_plot_gof}
\alias{stat_plot_gof}
\title{Plot the Goodness of fit of a (bt)ergm result}
\usage{
stat_plot_gof(gof)
}
\arguments{
\item{gof}{A gof model object for a fitted \code{ergm} or \code{btergm} model}
}
\value{
Plots displaying the Goodness of Fit
}
\description{
Plots the Goodness of fit for the output of a model of class \code{ergm} or \code{btergm}
}
\details{
This function plots the goodness-of-fit of unipartite Exponential Random Graphs Models 
or BTERGM models.

It does not process bipartite ERGMs model objects nor other types of 
Exponential random graph models object.
}
\examples{
\dontrun{
### ergm model
data(florentine, package = "snafun")
flom <- to_network(florentine$flomarriage)
m_ergm <- ergm::ergm(flom ~ edges + nodecov("Wealth"))
gof_ergm <- ergm::gof(m_ergm)
stat_plot_gof(gof_ergm)

### btergm model
data(alliances, package = "SNA4DSData")
m_btergm <- btergm::btergm(allianceNets ~ edges + 
  gwesp(0, fixed = TRUE) +
  edgecov(lastYearsSharedPartners) + 
  edgecov(militaryDisputes) + 
  nodecov("polity") +
  nodecov("Composite_Index_National_Capability") + 
  absdiff("polity") + 
  absdiff("Composite_Index_National_Capability") +
  edgecov(sharedBorder), 
  R = 100, # number of bootstraps
  parallel = "snow", ncpus = 16  # optional line
)

gof_btergm <- btergm:::gof.btergm(m_btergm)
stat_plot_gof(gof_btergm)
}
}
\seealso{
Other statistics functions: 
\code{\link{stat_ef_int}()},
\code{\link{stat_plot_gof_as_btergm}()}
}
\concept{statistics functions}
