% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_qap_lm.R
\name{summary.stat_qap_lm}
\alias{summary.stat_qap_lm}
\title{Summarize a QAP linear model result}
\usage{
\method{summary}{stat_qap_lm}(object, ...)
}
\arguments{
\item{object}{Object returned by \code{\link{stat_qap_lm}}.}

\item{...}{Ignored.}
}
\value{
An object of class \code{summary.stat_qap_lm}.
}
\description{
Summarize the coefficient-level results from \code{\link{stat_qap_lm}}.
}
