% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_matrix.R
\name{to_matrix}
\alias{to_matrix}
\title{Make a matrix from a graph}
\usage{
to_matrix(x)
}
\arguments{
\item{x}{graph object}
}
\value{
a \code{matrix} object
}
\description{
Make an matrix from various input graph types.
}
\details{
Create the appropriate matrix from various inputs:

\describe{
\item{igraph}{a bipartite network is turned into an incidence matrix, 
any other network becomes an adjacency matrix}
\item{network}{a bipartite network is turned into an incidence matrix, 
any other network becomes an adjacency matrix (non-sparse)}
\item{data.frame}{the first two columns are taken as, respectively, denoting 
'from' and 'to'. The third column is then used for the values of the cells 
in the matrix.}
}

If the input graph is bipartite, the resulting matrix will be an incidence 
matrix, otherwise it will be an adjacency natrix (weighted if the original 
graph is weighted).
}
\examples{
actors <- data.frame(name=c("Alice", "Bob", "Cecil", "David", "Esmeralda"), 
   age=c(48,33,45,34,21),
   gender=c("F","M","F","M","F"))
relations <- data.frame(from=c("Bob", "Cecil", "Cecil", "David", 
   "David", "Esmeralda"),
   to = c("Alice", "Bob", "Alice", "Alice", "Bob", "Alice"), 
   same.dept = c(FALSE, FALSE, TRUE, FALSE, FALSE, TRUE), 
   friendship = c(4, 5, 5, 2, 1, 1), 
   advice = c(4, 5, 5, 4, 2, 3))
}
